server/requirements.txt
python-decouple==3.3
IMAPClient==2.1.0